# Speed up tile building: don't copy GraphTileBuilder every time