## Issue

### What issue is this PR targeting? If there is no issue that addresses the problem, please open a corresponding issue and link it here.

Profiling `valhalla_build_tiles` target on my machine(mac OS) showed that `GraphTileBuilder` copy-construction and destruction take significant amount of time. Here is stacktrace from OS X "Time Profiler" 
<img width="884" alt="Screen Shot 2020-07-13 at 22 06 12" src="https://user-images.githubusercontent.com/14297128/87343288-48a6bb80-c555-11ea-8b76-4d91c84402f5.png">

The further analysis brought me to the `IsNextEdgeInternal` function. As you see the first line of the function copies `GraphTileBuilder`. https://github.com/valhalla/valhalla/blob/2ac1fa009c305e6f472c032e89f938e59c25cbe1/src/mjolnir/graphenhancer.cc#L884

The next experiments showed 10+% performance improvements:
| Source osm pbf  | pbf size | time without fix | time with fix | delta % |
| ------------- | ------------- | ------------- | ------------- | ------------- |
| belarus | 210 Mb  | 4:18 | 3:32 | 17.8% |
| germany | 3.4 Gb | ~1:36:00 | ~1:21:00 | 15.6% |

UPD:
the default valhalla config was used for experiments

Open questions:
1) Do we have regression tests to make sure that I didn't mess up the tile structure?
2) Do we need to benchmark the fix on larger amount of data? My expectations(hopes) that profit will be ~linear to the data size scale. However, It may depend on the number of threads/other specs of the server machine/config where the whole world tiles are built.

## Tasklist

 - [ ] Add tests
 - [ ] Add #fixes with the issue number that this PR addresses
 - [ ] Generally use squash merge to rebase and clean comments before merging
 - [ ] Update the [changelog](CHANGELOG.md)

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?