@gknisely 2020-07-15
> 1.  Do we have regression tests to make sure that I didn't mess up the tile structure?
> 2.  Do we need to benchmark the fix on larger amount of data? My expectations(hopes) that profit will be ~linear to the data size scale. However, It may depend on the number of threads/other specs of the server machine/config where the whole world tiles are built.

We should test this out in the test environment and RAD test it.