@kevinkreiser 2020-08-03
now that tile builds are stable, you can do a before and after checksum of the tiles created with utrecht for example to prove that this is safe, without RAD imho