@merkispavel 2020-08-04
> now that tile builds are stable, you can do a before and after checksum of the tiles created with utrecht for example to prove that this is safe, without RAD imho. maybe just in case use a larger pbf like switzerland or pennsylvania or some bigger geographical area.

oh, god lord. Tried to compare utrecht tiles hashsums - different. Did tile-building twice with master - level 2 tile has flaky hashsum :( Looking for the cause at the moment.

UPD: looks like edge bins are different. Guess: maybe edges are stored in different order