@kevinkreiser 2020-08-04
@merkispavel yeah i can see why that is happening. we do a random shuffle on the tilequeue so different threads get different tiles to bin. there are two things that get added to the bins, the edges that originate in the tile, i think those are sorted so should be a problem and then edges that cross through the tile but dont originate there (tweeners). those results come from different threads and are added on to the bins at the end. since the tiles are shuffled randomly this would cause different tweeners to come from different threads but we iterate through the threads in a fixed order at the end. this means the tweeners are in a randomized order which means we can do 1 of 2 things. pick a constant seed for the random shuffle, or sort the tweeners before we append them to the bins. the seed is probably the easiest but it might be nice (although dubiously useful) to have a more "sensible" ordering of them.