@merkispavel 2020-08-07
I've tested these changes with https://github.com/valhalla/valhalla/pull/2515: utrecht tiles checksums do match.
I guess the last question here is whether tests should be added to increase code coverage