@kevinkreiser 2020-08-07
@merkispavel i think we always want to be increasing test coverage, but since you didnt actually add anything new here i dont think adding coverage should really be done in this pr. to me we should look at the report and focus on the places with least coverage over time. i wouldnt block this pr over the coverage check and can merge anyway when we are ready to go