@kevinkreiser 2020-08-19
@gknisely pavel ran this on a larger tile set (258 tiles in PA) with parallelism turned on and confirmed that the output is the same (checksums of tiles in directories match with master)